pyproject.toml
[tool.poetry]
name = "tweet-stream"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
requests-oauthlib = "^1.3.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
